+++ packages/bitswap/test/stats.spec.ts
@@ -1,9 +1,10 @@
 import { generateKeyPair } from '@libp2p/crypto/keys'
 import { peerIdFromPrivateKey } from '@libp2p/peer-id'
 import { expect } from 'aegir/chai'
-import { stubInterface, type StubbedInstance } from 'sinon-ts'
+import { stubInterface } from 'sinon-ts'
 import { Stats } from '../src/stats.js'
 import type { Libp2p, MetricGroup, Metrics } from '@libp2p/interface'
+import type { StubbedInstance } from 'sinon-ts'
 
 interface StubbedStatsComponents {
   libp2p: StubbedInstance<Libp2p>