+++ packages/block-brokers/src/trustless-gateway/broker.ts
@@ -57,7 +57,7 @@ export class TrustlessGatewayBlockBroker implements BlockBroker<TrustlessGateway
       this.log('getting block for %c from %s', cid, gateway.url)
 
       try {
-        const block = await gateway.getRawBlock(cid, options.signal)
+        const block = await gateway.getRawBlock(cid, options)
         this.log.trace('got block for %c from %s', cid, gateway.url)
 
         try {