+++ packages/interop/src/fixtures/create-kubo.ts
@@ -1,6 +1,7 @@
-import { type KuboNode, createNode } from 'ipfsd-ctl'
+import { createNode } from 'ipfsd-ctl'
 import { path as kuboPath } from 'kubo'
 import { create as kuboRpcClient } from 'kubo-rpc-client'
+import type { KuboNode } from 'ipfsd-ctl'
 
 export async function createKuboNode (): Promise<KuboNode> {
   return createNode({