+++ packages/ipns/src/routing/helia.ts
@@ -1,7 +1,8 @@
-import { CustomProgressEvent, type ProgressEvent } from 'progress-events'
+import { CustomProgressEvent } from 'progress-events'
 import type { GetOptions, PutOptions } from './index.js'
 import type { IPNSRouting } from '../index.js'
 import type { Routing } from '@helia/interface'
+import type { ProgressEvent } from 'progress-events'
 
 export interface HeliaRoutingComponents {
   routing: Routing
@@ -22,6 +23,7 @@ export class HeliaRouting implements IPNSRouting {
       await this.routing.put(routingKey, marshaledRecord, options)
     } catch (err: any) {
       options.onProgress?.(new CustomProgressEvent<Error>('ipns:routing:helia:error', err))
+      throw err
     }
   }
 
@@ -30,9 +32,8 @@ export class HeliaRouting implements IPNSRouting {
       return await this.routing.get(routingKey, options)
     } catch (err: any) {
       options.onProgress?.(new CustomProgressEvent<Error>('ipns:routing:helia:error', err))
+      throw err
     }
-
-    throw new Error('Not found')
   }
 }