+++ packages/ipns/src/routing/local-store.ts
@@ -1,11 +1,13 @@
 import { Record } from '@libp2p/kad-dht'
-import { type Datastore, Key } from 'interface-datastore'
-import { CustomProgressEvent, type ProgressEvent } from 'progress-events'
+import { Key } from 'interface-datastore'
+import { CustomProgressEvent } from 'progress-events'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
-import type { GetOptions, PutOptions } from '../routing'
+import type { GetOptions, PutOptions } from '../routing/index.js'
 import type { AbortOptions } from '@libp2p/interface'
+import type { Datastore } from 'interface-datastore'
+import type { ProgressEvent } from 'progress-events'
 
 const DHT_RECORD_PREFIX = '/dht/record/'