+++ packages/ipns/test/publish.spec.ts
@@ -27,6 +27,7 @@ describe('publish', () => {
     customRouting = stubInterface<IPNSRouting>()
     customRouting.get.throws(new Error('Not found'))
     heliaRouting = stubInterface<Routing>()
+    dns = stubInterface<DNS>()
 
     name = ipns({
       datastore,
@@ -58,10 +59,15 @@ describe('publish', () => {
 
     expect(ipnsEntry).to.have.property('sequence', 1n)
 
-    // Ignore the milliseconds after the dot 2025-01-22T12:07:33.650000000Z
-    const expectedValidity = new Date(Date.now() + lifetime).toISOString().split('.')[0]
+    // Calculate expected validity as a Date object
+    const expectedValidity = new Date(Date.now() + lifetime)
 
-    expect(ipnsEntry.validity.split('.')[0]).to.equal(expectedValidity)
+    const actualValidity = new Date(ipnsEntry.validity)
+
+    const timeDifference = Math.abs(actualValidity.getTime() - expectedValidity.getTime())
+
+    // Allow a tolerance of 1 second (1000 milliseconds)
+    expect(timeDifference).to.be.lessThan(1000)
 
     expect(heliaRouting.put.called).to.be.true()
     expect(customRouting.put.called).to.be.true()